refactor: move new scoring to separate file
